+++ README.md
@@ -24,11 +24,11 @@ pip install veropt
 
 ## Usage
 
-Below is a simple example of running an optimisation problem with veropt. 
+Below is a simple example of running an optimisation problem with veropt.
 
 ```python
 from veropt import BayesOptimiser
-from veropt.obj_funcs.test_functions import *
+from veropt.obj_funcs.predefined_test_functions import *
 from veropt.gui import veropt_gui
 
 n_init_points = 24
@@ -36,7 +36,6 @@ n_bayes_points = 64
 n_evals_per_step = 4
 
 obj_func = PredefinedTestFunction("Hartmann")
-
 
 optimiser = BayesOptimiser(n_init_points, n_bayes_points, obj_func, n_evals_per_step=n_evals_per_step)
