+++ veropt/obj_funcs/__init__.py
@@ -1,14 +1,14 @@
 
 from veropt.obj_funcs import ocean_sims
-from veropt.obj_funcs import test_functions
+from veropt.obj_funcs import predefined_test_functions
 
 from veropt.obj_funcs.ocean_sims import OceanObjFunction
 
-from veropt.obj_funcs.test_functions import FitTestFunction, PredefinedTestFunction, PredefinedFitTestFunction
+from veropt.obj_funcs.predefined_test_functions import FitTestFunction, PredefinedTestFunction, PredefinedFitTestFunction
 
 __all__ = [
     "ocean_sims",
-    "test_functions",
+    "predefined_test_functions.py",
     "OceanObjFunction",
     "FitTestFunction",
     "PredefinedTestFunction",
